supabase/requirements.txt
streamlit
supabase
python-dotenv
sendgrid
